requirements.txt
praw==7.8.1
Flask==3.1.0
PyGithub==2.6.1
python-dotenv==1.0.1
